Reorganize sources and start work on FileRecordStore
