--- src/commonTest/kotlin/sample/SampleTests.kt
@@ -1,11 +1,0 @@
-package sample
-
-import kotlin.test.Test
-import kotlin.test.assertTrue
-
-class SampleTests {
-    @Test
-    fun testMe() {
-        assertTrue(Sample().checkMe() > 0)
-    }
-}
