+++ src/iosMain/kotlin/com/gu/ophan/FileRecordStore.kt
@@ -1,0 +1,17 @@
+package com.gu.ophan
+
+import com.gu.ophan.RecordStore
+
+actual class FileRecordStore actual constructor(path: String) : RecordStore {
+    override fun putRecord(key: String, record: ByteArray) {
+        TODO("not implemented") //To change body of created functions use File | Settings | File Templates.
+    }
+
+    override fun getRecords(): List<ByteArray> {
+        TODO("not implemented") //To change body of created functions use File | Settings | File Templates.
+    }
+
+    override fun removeRecord(key: String) {
+        TODO("not implemented") //To change body of created functions use File | Settings | File Templates.
+    }
+}
