+++ src/iosMain/kotlin/com/gu/ophan/MainQueueDispatcher.kt
@@ -1,8 +1,10 @@
-package sample
+package com.gu.ophan
 
-import kotlinx.coroutines.*
-import platform.darwin.*
-import kotlin.coroutines.*
+import kotlinx.coroutines.CoroutineDispatcher
+import kotlinx.coroutines.Runnable
+import platform.darwin.dispatch_async
+import platform.darwin.dispatch_get_main_queue
+import kotlin.coroutines.CoroutineContext
 
 class MainQueueDispatcher : CoroutineDispatcher() {
     override fun dispatch(context: CoroutineContext, block: Runnable) {
