--- src/iosMain/kotlin/sample/SampleIos.kt
@@ -1,9 +1,0 @@
-package sample
-
-actual class Sample {
-    actual fun checkMe() = 7
-}
-
-actual object Platform {
-    actual fun name(): String = "iOS"
-}
