--- src/iosTest/kotlin/sample/SampleTestsNative.kt
@@ -1,11 +1,0 @@
-package sample
-
-import kotlin.test.Test
-import kotlin.test.assertTrue
-
-class SampleTestsNative {
-    @Test
-    fun testHello() {
-        assertTrue("iOS" in hello())
-    }
-}
