--- src/jvmMain/kotlin/sample/SampleJvm.kt
@@ -1,9 +1,0 @@
-package sample
-
-actual class Sample {
-    actual fun checkMe() = 42
-}
-
-actual object Platform {
-    actual fun name(): String = "JVM"
-}
